+++ docs/source/conf.py
@@ -86,9 +86,7 @@ TMP_FOLDER.mkdir(exist_ok=True)
 # zip the assets directory found in the examples directory and place it in the current dir
 def zip_dir(source_dir, zip_file_name):
     # Create archive with compressed files
-    with zipfile.ZipFile(
-        file=TMP_FOLDER / zip_file_name, mode="w", compression=zipfile.ZIP_DEFLATED
-    ) as ziph:
+    with zipfile.ZipFile(file=TMP_FOLDER / zip_file_name, mode="w", compression=zipfile.ZIP_DEFLATED) as ziph:
         for root, _, files in os.walk(source_dir):
             for file in files:
                 ziph.write(
@@ -203,6 +201,7 @@ for repo in SUBSTRA_REPOS:
         source_path = Path(imported_module.__file__).resolve().parents[1] / repo.doc_dir
         copy_source_files(source_path, repo.dest_doc_dir)
 
+
 # reformat links to a section in a markdown files (not supported by myst_parser)
 def reformat_md_section_links(file_path: Path):
     # Read in the file
@@ -312,8 +311,8 @@ nitpick_ignore = [
     ("py:class", "substra.Client"),
     ("py:class", "substra.sdk.client.Client"),
     ("py:class", "substra.sdk.models.ComputePlan"),
-    ("py:class", "substra.sdk.schemas.AlgoOutputSpec"),
-    ("py:class", "substra.sdk.schemas.AlgoInputSpec"),
+    ("py:class", "substra.sdk.schemas.FunctionOutputSpec"),
+    ("py:class", "substra.sdk.schemas.FunctionInputSpec"),
     ("py:class", "ComputePlan"),
 ]
 
@@ -392,9 +391,7 @@ sphinx_gallery_conf = {
         "repo": "substra-documentation",
         "branch": current_commit,  # Can be any branch, tag, or commit hash. Use a branch that hosts your docs.
         "binderhub_url": "https://mybinder.org",  # public binderhub url
-        "dependencies": str(
-            Path(__file__).parents[2] / "requirements.txt"
-        ),  # this value is not used
+        "dependencies": str(Path(__file__).parents[2] / "requirements.txt"),  # this value is not used
         "notebooks_dir": "notebooks",
         "use_jupyter_lab": True,
     },